Filter instances by bastion VPC
If you're SSH'ing to a bastion, probably you only want to see machines
which are in the same VPC. This means you don't get machines listed
which aren't reachable.

In the future we might want to make this somehow configurable, but I
don't have a use case for that yet.